Render static routes without suspension
