+++ src/client/App.tsx
@@ -1,93 +1,85 @@
-import { Suspense } from "react";
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import AuthenticatedRoute from "@/components/account/authenticated-route";
+import CookiePolicy from "@/components/legal/cookie-policy";
+import DataDeletion from "@/components/legal/data-deletion";
+import LegalIndex from "@/components/legal/legal-index";
+import PrivacyPolicy from "@/components/legal/privacy-policy";
+import TermsOfService from "@/components/legal/tos";
 import SiteMeta from "@/components/site-meta";
-import { RouteFallback } from "@/components/layout/route-fallback";
+import AccountPrivacyPage from "./pages/AccountPrivacy";
+import BotPage from "./pages/Bot";
+import CompanionPage from "./pages/Companion";
 import HomePage from "./pages/Home";
 import LegalPage from "./pages/LegalPage";
 import LoginPage from "./pages/Login";
+import MapAnalyzerPage from "./pages/MapAnalyzer";
+import NotFoundPage from "./pages/NotFound";
+import OsuGuessrPage from "./pages/OsuGuessr";
 import ProfilePage from "./pages/Profile";
 import { legacyRedirects, routes } from "./routes/paths";
-import { routeModules } from "./routes/route-modules";
-
-const {
-    accountPrivacy,
-    bot,
-    companion,
-    legalCookies,
-    legalDataDeletion,
-    legalIndex,
-    legalPrivacy,
-    legalTerms,
-    mapAnalyzer,
-    notFound,
-    osuguessr,
-} = routeModules;
 
 export default function App() {
     return (
         <BrowserRouter>
             <SiteMeta />
             <div className="min-h-screen">
-                <Suspense fallback={<RouteFallback />}>
-                    <Routes>
-                        <Route path={routes.home} element={<HomePage />} />
-                        <Route path={routes.bot} element={<bot.Component />} />
-                        <Route path={routes.osuguessr} element={<osuguessr.Component />} />
-                        <Route path={routes.companion} element={<companion.Component />} />
-                        <Route path={routes.mapAnalyzer} element={<mapAnalyzer.Component />} />
-                        <Route
-                            path={routes.legal}
-                            element={
-                                <LegalPage>
-                                    <legalIndex.Component />
-                                </LegalPage>
-                            }
-                        />
-                        <Route
-                            path={routes.legalPrivacy}
-                            element={
-                                <LegalPage>
-                                    <legalPrivacy.Component />
-                                </LegalPage>
-                            }
-                        />
-                        <Route
-                            path={routes.legalTerms}
-                            element={
-                                <LegalPage>
-                                    <legalTerms.Component />
-                                </LegalPage>
-                            }
-                        />
-                        <Route
-                            path={routes.legalCookies}
-                            element={
-                                <LegalPage>
-                                    <legalCookies.Component />
-                                </LegalPage>
-                            }
-                        />
-                        <Route
-                            path={routes.legalDataDeletion}
-                            element={
-                                <LegalPage>
-                                    <legalDataDeletion.Component />
-                                </LegalPage>
-                            }
-                        />
-                        {Object.entries(legacyRedirects).map(([path, destination]) => (
-                            <Route key={path} path={path} element={<Navigate to={destination} replace />} />
-                        ))}
-                        <Route element={<AuthenticatedRoute />}>
-                            <Route path={routes.profile} element={<ProfilePage />} />
-                            <Route path={routes.profilePrivacy} element={<accountPrivacy.Component />} />
-                            <Route path={routes.profilePrivacyConfirm} element={<accountPrivacy.Component />} />
-                        </Route>
-                        <Route path={routes.login} element={<LoginPage />} />
-                        <Route path="*" element={<notFound.Component />} />
-                    </Routes>
-                </Suspense>
+                <Routes>
+                    <Route path={routes.home} element={<HomePage />} />
+                    <Route path={routes.bot} element={<BotPage />} />
+                    <Route path={routes.osuguessr} element={<OsuGuessrPage />} />
+                    <Route path={routes.companion} element={<CompanionPage />} />
+                    <Route path={routes.mapAnalyzer} element={<MapAnalyzerPage />} />
+                    <Route
+                        path={routes.legal}
+                        element={
+                            <LegalPage>
+                                <LegalIndex />
+                            </LegalPage>
+                        }
+                    />
+                    <Route
+                        path={routes.legalPrivacy}
+                        element={
+                            <LegalPage>
+                                <PrivacyPolicy />
+                            </LegalPage>
+                        }
+                    />
+                    <Route
+                        path={routes.legalTerms}
+                        element={
+                            <LegalPage>
+                                <TermsOfService />
+                            </LegalPage>
+                        }
+                    />
+                    <Route
+                        path={routes.legalCookies}
+                        element={
+                            <LegalPage>
+                                <CookiePolicy />
+                            </LegalPage>
+                        }
+                    />
+                    <Route
+                        path={routes.legalDataDeletion}
+                        element={
+                            <LegalPage>
+                                <DataDeletion />
+                            </LegalPage>
+                        }
+                    />
+                    {Object.entries(legacyRedirects).map(([path, destination]) => (
+                        <Route key={path} path={path} element={<Navigate to={destination} replace />} />
+                    ))}
+                    <Route element={<AuthenticatedRoute />}>
+                        <Route path={routes.profile} element={<ProfilePage />} />
+                        <Route path={routes.profilePrivacy} element={<AccountPrivacyPage />} />
+                        <Route path={routes.profilePrivacyConfirm} element={<AccountPrivacyPage />} />
+                    </Route>
+                    <Route path={routes.login} element={<LoginPage />} />
+                    <Route path="*" element={<NotFoundPage />} />
+                </Routes>
             </div>
         </BrowserRouter>
     );
