+++ src/client/pages/NotFound.tsx
@@ -9,12 +9,10 @@ export default function NotFound() {
         <div className={sitePageClass}>
             <Header />
             <main className={cn(siteContainerClass, "flex min-h-155 flex-col items-start justify-center py-20")}>
-                <span className="mb-[1.3rem] font-mono text-[0.75rem] text-quiet">404</span>
                 <Eyebrow>Lost in the mapset</Eyebrow>
-                <h1 className="max-w-180 text-[clamp(2.8rem,7vw,5.4rem)] leading-none tracking-[-0.06em]">This route does not exist.</h1>
-                <p className="mt-[1.3rem] mb-8 max-w-[52ch] leading-[1.7] text-muted">
-                    The ecosystem is still here. Head back to the project index and choose another path.
-                </p>
+                <h1 className="mb-8 max-w-180 text-[clamp(2.8rem,7vw,5.4rem)] leading-none tracking-[-0.06em]">
+                    This route does not exist.
+                </h1>
                 <ActionLink href="/">Return home</ActionLink>
             </main>
             <Footer />
